Add files via upload
This small library will allow you to create a sequence of numbers and scramble them. For usage, you add a reference to the DLL, or the Vriliancy_Digits class, and call the CreateNumbers method, and pass in integer values for your starting point, and your ending point. The script will generate numbers between and including the starting point/ending point numbers. Note that this method returns a Queue<Tuple<int,int>>